requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
ase==3.25.0
async-timeout==5.0.1
attrs==25.3.0
bibtexparser==1.4.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
docker-pycreds==0.4.0
filelock==3.13.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
joblib==1.5.0
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==2.1.5
matplotlib==3.10.3
monty==2025.3.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.18
narwhals==1.39.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
p_tqdm==1.4.2
packaging==25.0
palettable==3.3.3
pandarallel==1.6.5
pandas==2.2.3
pathos==0.3.4
pillow==11.0.0
platformdirs==4.3.8
plotly==6.0.1
pox==0.3.6
ppft==1.7.7
propcache==0.3.1
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pymatgen==2025.5.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytorch-lightning==2.1.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.3
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
SMACT==3.1.0
smmap==5.0.2
spglib==2.6.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
torch==2.3.0+cu121
torch-geometric==2.6.1
torch_scatter==2.1.2+pt23cu121
torch_sparse==0.6.18+pt23cu121
torchaudio==2.3.0+cu121
torchmetrics==1.7.1
torchvision==0.18.0+cu121
tqdm==4.67.1
triton==2.3.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
uncertainties==3.2.3
urllib3==2.4.0
wandb==0.19.11
yarl==1.20.0
